+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- VIMEET-2261 - Add notification new request count
+
+
 ## [2.71.4] - 2020-11-30
 ### Fixed
 - MV-281 - Fix webinar recording layout when HLS is activated